Creacion del nodo base

+++ Criaturas.cpp
@@ -1,6 +1,8 @@
 #include "Criaturas.h"
 
-Criaturas::Criaturas(const string &nombre, int vida, int edad, int x, int y): nombre(nombre), vida(vida), edad(edad), posicion(make_pair(x, y)) {}
+Criaturas::Criaturas(const string &nombre, int vida, int edad, int x, int y): nombre(nombre), vida(vida), edad(edad),
+                                                                              posicion(make_pair(x, y)) {
+}
 
 string Criaturas::getNombre() {
     return nombre;
@@ -14,7 +16,7 @@ int Criaturas::getEdad() {
     return edad;
 }
 
-pair<float, float> Criaturas::getPosicion() {
+pair<int, int> Criaturas::getPosicion() {
     return posicion;
 }
